check last 5 pages of thread (rct tool)
in the event a new recruit's post is not on the last page (in the case
of clanwide threads, which change by the minute), recursively check the
next to last page until a match is found. Max of five tries, to prevent
traffic load. This should never fail if the user posted that day.

closes #39